requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Session==0.5.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
Pillow==9.5.0
polling==0.3.2
pycodestyle==2.10.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
quickdraw==1.0.0
redis==4.5.5
requests==2.30.0
six==1.16.0
tomli==2.0.1
twilio==8.2.0
urllib3==2.0.2
Werkzeug==2.3.4
yarl==1.9.2
